+++ Changelog.md
@@ -25,6 +25,7 @@
   - Will now use rendered mode without overhang detection, the default fov, and the correct aspect ratio
 - Pick between Perspective or Orthographic camera projections
 - Camera pitch is no longer locked between (-90°, 90°)
+- Exposure test generator ([#24](https://github.com/connorslade/mslicer/pull/24))
 
 ## v0.5.0 &mdash; February 21st, 2026